requirements.txt
python-binance==0.7.9
pandas==1.2.3 